requirements.txt
duckduckgo-search
groq
python-dotenv
